+++ guiclient/todoListCalendar.cpp
@@ -390,7 +390,7 @@ void todoListCalendar::resizeEvent(QResizeEvent* event)
 {
   XWidget::resizeEvent(event);
 
-  _gview->setMinimumWidth(_gview->height() * (_gview->scene()->sceneRect().width() / _gview->scene()->sceneRect().height()));
+  _gview->setMinimumWidth((int)(_gview->height() * (_gview->scene()->sceneRect().width() / _gview->scene()->sceneRect().height())));
   _gview->fitInView(_gview->scene()->sceneRect(), Qt::KeepAspectRatio);
 }
 
@@ -398,6 +398,6 @@ void todoListCalendar::showEvent(QShowEvent * event)
 {
   XWidget::showEvent(event);
 
-  _gview->setMinimumWidth(_gview->height() * (_gview->scene()->sceneRect().width() / _gview->scene()->sceneRect().height()));
+  _gview->setMinimumWidth((int)(_gview->height() * (_gview->scene()->sceneRect().width() / _gview->scene()->sceneRect().height())));
   _gview->fitInView(_gview->scene()->sceneRect(), Qt::KeepAspectRatio);
 }
